+++ CHANGELOG.md
@@ -13,6 +13,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - `persp-ibuffer-generate-filter-groups`: load `ibuf-ext` library if needed ([#202](https://github.com/nex3/perspective-el/issues/202)).
 
 
+### Added
+
+- `persp-kill-other-buffers`: kill other buffers in the current perspectives.
+
+
 ## [2.18] — 2022-09-20
 
 ### Added